code: import reservations and reservation requests
This commits adds the DB and the migration for the reservations data.
The docker-compose based stack is updated to  prepare the environment
for development or testing to wait for the DB to be ready before
the execution starts.

The tests had to be updated since we moved from sync commands to Promise
based.

comments